making utils service
