+++ src/app/photo-rec/photo-rec.page.ts
@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { PhotoService } from '../services/photo.service';
 import { Router } from '@angular/router';
 import { ApiService } from '../services/api.service';
-import { ToastController, ModalController } from '@ionic/angular';
+import { UtilsService } from '../services/utils.service';
+import { PreviewComponent } from '../preview/preview.component';
 
 @Component({
   selector: 'app-photo-rec',
@@ -16,8 +17,7 @@ export class PhotoRecognize implements OnInit {
     public photoService: PhotoService,
     public router: Router,
     public apiService: ApiService,
-    public toastController: ToastController,
-    public modalController: ModalController
+    public utils: UtilsService
   ) { }
 
   ngOnInit() {
@@ -26,7 +26,7 @@ export class PhotoRecognize implements OnInit {
 
   open(photo) {
     this.currentImage = photo;
-    this.presentModal();
+    this.utils.presentModal(PreviewComponent);
 
   }
 
@@ -42,25 +42,12 @@ export class PhotoRecognize implements OnInit {
     this.apiService.uploadLoteOfPhotos(this.photoService.photos).subscribe(data => {
       console.log(data);
       this.clearPhotos();
-      this.presentToast('Processing performed successfully. We will send you an email. 🎉');
+      this.utils.presentToast('Processing performed successfully. We will send you an email. 🎉');
     }, err => {
       console.log(err);
-      this.presentToast('We had an error uploading, please try again! 🥺');
+      this.utils.presentToast('We had an error uploading, please try again! 🥺');
     });
   }
 
-  async presentToast(msg: string) {
-    const toast = await this.toastController.create({
-      message: msg,
-      duration: 2000
-    });
-    toast.present();
-  }
 
-  async presentModal() {
-/*     const modal = await this.modalController.create({
-      component: 
-    });
-    return await modal.present(); */
-  }
 }
